fix: complete packet similarity threshold tuning and test data seeding

+++ SWD-Grading/BLL/Interface/IPacketSimilarityTestDataSeeder.cs
@@ -8,6 +8,6 @@ namespace BLL.Interface
 {
     public interface IPacketSimilarityTestDataSeeder
     {
-        Task<string> SeedAsync(int examId);
+        Task<string> SeedAsync(long examId);
     }
 }
